+++ packages/core/src/identity/index.ts
@@ -1,0 +1,4 @@
+export type * from "./model.js";
+export { IDENTITY_NAMESPACES, normalizeIdentityId, resolveIdentityClaims } from "./model.js";
+export type { IdentityResolverSource, IdentityResolverOptions } from "./resolver.js";
+export { IdentityResolver } from "./resolver.js";